mps-server project now links directly to the wisock library
This is necessary only if gcc (i.e., gmake) is the target compiler.
GCC does not include liked libraries when generating a static lib.
Since aerolib.lib links to ws2_32, when built with vs2022 it will
automatically inlcude ws2_32 in aerolib.lib, but when built with gcc
it will not.